Rework moving_average to make it more easily threadsafe. Update lots of tests. Disable regular_updater_test ~ as it was currently relying on sketchy performance characteristics & attempts to mock these things were defeated by vc13's poor stdlib implementation.

+++ test/snapshot_test.cpp
@@ -2,7 +2,7 @@
 
 BOOST_AUTO_TEST_CASE(snapshot_test){
 	std::vector<int> vec(17);
-	for(int i = 0; i < vec.size(); ++i){
+	for(int i = 0; i < (int)vec.size(); ++i){
 		vec[i] = (32-i)*(i+18) + 2;
 	}
